Finish string namespace
